Refactor some code to untighten the coupling between the identifier and the ExtensionAdapter
